Safari-611.1.10/LayoutTests/fast/hidpi/image-srcset-encoded.html

<html>
    <head>
        <script src="../../resources/js-test-pre.js"></script>
        <script>
             addEventListener("load", function() {
                 shouldBeTrue('document.getElementById("foo").clientWidth==25');
             }, false);
        </script>
    </head>

    <body id="body">
        <div>This test passes if the image below is displayed.</div>
        <img id="foo" srcset="resources/image%40test.png">
    </body>
</html>
